[Security Solution] Move Endpoint details flyout and related middleware to their own functions (#108330)
* Split function to load endpoint details using parameters

* Add action to load endpoint details

* Moving all the endpoint details content to a separate component

* Rename endpoint details to endpoint details content

* Rename temporal file into EndpointDetails

* Remove unused dispatch

* Refactor ingestPolicies dispatching in the middleware
